fix test file names
